backend/requirements-test.txt
pytest
pytest_django
pytest_random
hypothesis
model_bakery
django_core